Refactor calls to CVideo::flip()
This adds a timer in the SDL event queue that fires every 20ms in order to add a draw event into the queue. After the draw event has been processed by all listener, a call to CVideo::flip() is made. All other calls to CVideo::flip() have been removed.

This is the first step towards simplifying the rendering path and making sure the screen is in a consistent state when the flip() call is made. Having the screen in a consistent state will resolve flickering.

This incidentially also fixes resizing in load screens by always blanking them and doing a full re-render.