Add a FPS counter that's invoked every time a draw call is made.
Change the codebase so that draws are now event-driven, and every time an event is pumped, there may be a draw performed. Draw-events have been turned into singletons, so there will never be more than one in the queue.
This code also disables the loadscreen since it'll conflict with master anyway. This is still work in progress, and not complete, but is a step towards getting there.